EC: Blur: Color Filters implementation (2/2)
- Since a part of the code was already available in the Blur base code,
let's fix the remaining stuff and use it.

How does it works?

- It uses the lightness level of the background in use to determine
which color to apply for the Blur bckground. So for example the
background is a more lighter one a different color will be used and so
on...
- If this cannot be determined, a third type of color filter will be
used which is like a sort of mixed combination of the other two  ones,
which kinda good looking.

There are three color filters:

- Light color filter (Used for Light backgrounds)
- Dark color filter (Used for Dark backgrounds)
- Mixed color filter (Used for Dark/light backgrounds)

- I made the colors to be applied simultaneously for both Blur
backgrounds: StatusBar and RecentApps ones.
- I took this decision because in my opinion there's no reason to make
three more different ColorPicker preferences just for Recents, if there
will be requests on that it will be implemented. But for now we stick to
this.

Change-Id: I11b1d2ed83c07c4be5050117c2e48b40bb652929
Signed-off-by: Calin <nilac8991@gmail.com>
Signed-off-by: Pafcholini <nadyaivanova14@gmail.com>